Refactor bn-trips and bn-hauls into libs/bn-common
