+++ libs/bn-common/.gitignore
