+++ libs/bn-common/babel.config.js
