+++ libs/bn-common/public/favicon.ico
