+++ libs/bn-common/public/index.html
@@ -5,11 +5,11 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width,initial-scale=1.0">
     <link rel="icon" href="<%= BASE_URL %>favicon.ico">
-    <title>bn-hauls</title>
+    <title>bn-common</title>
   </head>
   <body>
     <noscript>
-      <strong>We're sorry but bn-hauls doesn't work properly without JavaScript enabled. Please enable it to continue.</strong>
+      <strong>We're sorry but bn-common doesn't work properly without JavaScript enabled. Please enable it to continue.</strong>
     </noscript>
     <div id="app"></div>
     <!-- built files will be auto injected -->
