+++ libs/bn-common/src/models/BoatnetHaulsSettings.ts
