+++ libs/bn-common/src/models/BoatnetTripsSettings.ts
@@ -1,0 +1,3 @@
+export interface BoatnetTripsSettings {
+  name: string;
+}
