+++ libs/bn-common/src/shims-tsx.d.ts
