+++ libs/bn-common/src/shims-vue.d.ts
