+++ libs/bn-common/tsconfig.json
